remote_agents/teams-bot/requirements.txt
fastapi
uvicorn
botbuilder-core
botbuilder-schema
python-dotenv
aiohttp